pyproject.toml
[tool.poetry]
name = "langchain-agents"
version = "0.1.0"
description = ""
authors = ["Iwo Dziechciarow <iiwo@o2.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-jose = "^3.3.0"
pytest = "^7.4.0"
langchain = { version = "0.0.286", extras = ['openai'] }
langchain-experimental = "^0.0.16"
python-dotenv = "^1.0.0"
openai = "^0.28"
tiktoken = "^0.3.2"
duckduckgo-search = "^3.8.5"
chromadb = "^0.3.23"
pandas = "2.0.3"
playwright = "^1.37.0"
beautifulsoup4 = "^4.12.2"
faiss-cpu = "^1.7.4"
langchain-plantuml = "^0.0.7"
langchain-visualizer = "^0.0.29"
vcr-langchain = "^0.0.29"
ansi2html = "^1.8.0"
loguru = "^0.7.2"

[tool.poetry.extras]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
